Wrong variable type
git-svn-id: http://openmesh.org/svnrepo/OpenMesh@861 fdac6126-5c0c-442c-9429-916003d36597

+++ trunk/src/OpenMesh/Tools/Decimater/McDecimaterT.hh
@@ -126,7 +126,7 @@ private: //------------------------------------------------------- private data
   // reference to mesh
   Mesh&      mesh_;
 
-  unsigned int randomSamples_;
+  size_t randomSamples_;
 
 };
